Moved files into tests directory
